refactor: move the grammar parsing engine into homeboy-engine-primitives (#8425)
grammar (the tree-sitter-style structural parsing engine: ~3000 lines across
grammar/, grammar_items, grammar_strings) is a language-agnostic primitive with
zero built-in language knowledge — yet it was filed under extension/. Its entire
external footprint is homeboy_error (4x) + engine::local_files (1x, itself now a
primitive) + serde/toml/regex. Classic misplaced primitive, same pattern as
Language (#8433) but larger.

Move the whole cluster into homeboy-engine-primitives as a cohesive `grammar`
module (grammar_items -> grammar::items, grammar_strings -> grammar::strings):
- rewire the internal super:: refs to the new nesting
- crate::error -> homeboy_error; crate::engine::local_files -> crate::local_files
- add toml as a primitives dependency (grammar.toml loading)
- promote cached_regex pub(crate) -> pub for cross-crate access by core_fingerprint
- extension/mod.rs re-exports `grammar` (+ `grammar_items` alias) so the ~10
  consumers keep resolving crate::extension::grammar unchanged
- repoint engine::symbol_graph directly at the primitives grammar, severing an
  engine -> extension edge (9 -> 8)

Behavior-preserving. Verified: primitives + core compile clean; 44 grammar tests
pass.

Refs #8425
