+++ crates/homeboy-engine-primitives/Cargo.toml
@@ -17,6 +17,7 @@ homeboy-error = { version = "0.1.0", path = "../homeboy-error" }
 homeboy-paths = { version = "0.1.0", path = "../homeboy-paths" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+toml = "1.0.3"
 regex = "1.11"
 libc = "0.2"
